API/Wslny/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
gunicorn>=21.2
djangorestframework-simplejwt>=5.3
google-auth>=2.23
requests>=2.31
